+++ tests/test_unchecked_api.cpp
@@ -42,7 +42,7 @@ TEST(UnCheckedAPITests, test_next)
 {
     const char* twochars = "\xe6\x97\xa5\xd1\x88";
     const char* w = twochars;
-    int cp = utf8::unchecked::next(w);
+    unsigned int cp = utf8::unchecked::next(w);
     EXPECT_EQ (cp, 0x65e5);
     EXPECT_EQ (w, twochars + 3);
 
@@ -65,7 +65,7 @@ TEST(UnCheckedAPITests, test_next)
 TEST(UnCheckedAPITests, test_peek_next)
 {
     const char* const cw = "\xe6\x97\xa5\xd1\x88";
-    int cp = peek_next(cw);
+    unsigned int cp = peek_next(cw);
     EXPECT_EQ (cp, 0x65e5);
 }
 
@@ -73,7 +73,7 @@ TEST(UnCheckedAPITests, test_prior)
 {
     const char* twochars = "\xe6\x97\xa5\xd1\x88";
     const char* w = twochars + 3;
-    int cp = prior (w);
+    unsigned int cp = prior (w);
     EXPECT_EQ (cp, 0x65e5);
     EXPECT_EQ (w, twochars);
 
@@ -109,13 +109,13 @@ TEST(UnCheckedAPITests, test_advance)
 TEST(UnCheckedAPITests, test_distance)
 {
     const char* twochars = "\xe6\x97\xa5\xd1\x88";
-    size_t dist = utf8::unchecked::distance(twochars, twochars + 5);
+    size_t dist = static_cast<size_t>(utf8::unchecked::distance(twochars, twochars + 5));
     EXPECT_EQ (dist, 2);
 }
 
 TEST(UnCheckedAPITests, test_utf32to8)
 {
-    int utf32string[] = {0x448, 0x65E5, 0x10346, 0};
+    unsigned int utf32string[] = {0x448, 0x65E5, 0x10346, 0};
     string utf8result;
     utf32to8(utf32string, utf32string + 3, back_inserter(utf8result));
     EXPECT_EQ (utf8result.size(), 9);
@@ -124,7 +124,7 @@ TEST(UnCheckedAPITests, test_utf32to8)
 TEST(UnCheckedAPITests, test_utf8to32)
 {
     const char* twochars = "\xe6\x97\xa5\xd1\x88";
-    vector<int> utf32result;
+    vector<unsigned int> utf32result;
     utf8to32(twochars, twochars + 5, back_inserter(utf32result));
     EXPECT_EQ (utf32result.size(), 2);
 }
